scraper/requirements.txt
requests
beautifulsoup4
psycopg2-binary
clickhouse-connect
prefect
python-dotenv